WIP, convert the multiple time multiple domain.
This change compiles but does not properly work, probably because I'm
ignore the new parameters, `timestep` and `domain` to most of the routines.
I'll fix that shortly.

Synopsis:

Changed label, version and dbtype then in nn1stsd.xml

$ cd gen
$ xml2plugin ../nn1stsd.xml

I reverted all changes to build.make file and link.txt.

Using `git difftool avtnn1stsdFileFormat.*` I merged the changes in the
body of the existing routines I'd changed into the corresponding new
files. There are two new parameters in most of the files, timestep and
domain which, of course, are required for MTMD  and those I did not
change.

I accepted the changes verbatim to Makefile.cmake,
nn1stsdCommonPluginInfo.C and nn1stsdPluginInfo.C.